Moved fixture to separate class Added createStudent method and integration test for StudentApp
